Refactor analysis_view.py and split components into src/gui/analysis/
